we can get energy consumption

+++ src/main/java/nl/hr/cmi/inf/Dtos/Month.java
@@ -23,7 +23,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Month {
 
     @JsonProperty("timstamp")
-    private Integer timstamp;
+    private Long timstamp;
     @JsonProperty("value")
     private Integer value;
     @JsonProperty("pk")
@@ -39,7 +39,7 @@ public class Month {
      *     The timstamp
      */
     @JsonProperty("timstamp")
-    public Integer getTimstamp() {
+    public Long getTimstamp() {
         return timstamp;
     }
 
@@ -49,7 +49,7 @@ public class Month {
      *     The timstamp
      */
     @JsonProperty("timstamp")
-    public void setTimstamp(Integer timstamp) {
+    public void setTimstamp(Long timstamp) {
         this.timstamp = timstamp;
     }
